Update Practical-04-Reuse.qmd
In the section "The Advantages of a Package", practical 4, lines 355-357 in the code,
it prints: **The URL's data labels are: City, Population, Latitude, Longitude.**
This is incorrect, because it mixes with the previous variables with the same name. In this example, the url is different, and the text should say: **The URL's data labels are: City, Region, Founded, Population, URL, Longitude, Latitude**

I changed slightly the variable names in this code block to avoid overlap.